Fix js tests
[MAILPOET-2835]
