+++ assets/js/src/form_editor/blocks/divider/divider.ts
@@ -1,37 +1,7 @@
 import MailPoet from 'mailpoet';
 import icon from './icon.jsx';
 import edit from './edit';
-
-export enum Types {
-  Divider = 'divider',
-  Spacer = 'spacer',
-}
-
-export enum Style {
-  Solid = 'solid',
-  Dashed = 'dashed',
-  Dotted = 'dotted',
-}
-
-export interface Attributes {
-  className: string,
-  height: number;
-  type: Types;
-  style: Style;
-  dividerHeight: number;
-  dividerWidth: number;
-  color: string;
-}
-
-export const defaultAttributes = {
-  className: undefined,
-  height: 1,
-  type: Types.Divider,
-  style: Style.Solid,
-  dividerHeight: 1,
-  dividerWidth: 100,
-  color: 'black',
-};
+import { defaultAttributes } from './divider_types';
 
 export const name = 'mailpoet-form/divider';
